model_wrappers/gem/gem_repo/apps/molecular_docking/helixdock/requirements.txt
ml-collections
tensorboardX
pgl
rdkit-pypi==2022.3.3
paddlepaddle-gpu=2.4.1
tqdm
scikit-learn
